Refactor to Sender/Builder
