+++ tests/Chat/API/ChatSendTest.php
@@ -4,11 +4,20 @@ use Shipweb\LineConnect\Chat\API\ChatSend;
 use Shipweb\LineConnect\Core\LineConnect;
 use Shipweb\LineConnect\Message\LINE\Builder;
 
+require_once __DIR__ . '/../../../tests/LINEBot/Util/DummyHttpClient.php';
+
 class ChatSendTest extends WP_Ajax_UnitTestCase {
 
     public function setUp(): void {
         parent::setUp();
         add_action('wp_ajax_lc_ajax_chat_send', [ChatSend::class, 'ajax_chat_send']);
+        add_filter(LineConnect::FILTER_PREFIX . 'httpclient', function ($httpClient) {
+            $mock = function ($testRunner, $httpMethod, $url, $data) {
+                return ['status' => 200];
+            };
+            $dummyHttpClient = new LINE\Tests\LINEBot\Util\DummyHttpClient($this, $mock);
+            return $dummyHttpClient;
+        });
         lineconnectTest::init();
     }
 
@@ -17,25 +26,25 @@ class ChatSendTest extends WP_Ajax_UnitTestCase {
             'action' => 'lc_ajax_chat_send',
             'nonce' => wp_create_nonce(LineConnect::CREDENTIAL_ACTION__POST),
             'channel' => '04f7',
-            'to' => 'Ud2be13c6f39c97f05c683d92c696483a',
+            'to' => 'U4d3e3f7cf0c84c18526ee3f2c8d9d543', // none existing user id
             'messages' => $messages,
             'notificationDisabled' => 0
         ];
     }
 
-    public function test_failed_message_send() {
+    public function test_success_message_send() {
         $this->_setRole('administrator');
         $_POST = $this->create_valid_post_data(
             [
-                [ 'type' => 'text'],
-                [ 'message' => ['text' => [ 'text' => 'Hello']]]
+                ['type' => 'text'],
+                ['message' => ['text' => ['text' => 'Hello']]]
             ]
         );
 
         $response = $this->handle_ajax_request();
-
-        $this->assertFalse($response['success']);
-        $this->assertEquals('failed', $response['data']['result']);
+        var_dump($response);
+        $this->assertTrue($response['success']);
+        $this->assertEquals('success', $response['data']['result']);
     }
 
     public function test_invalid_nonce() {
@@ -53,14 +62,14 @@ class ChatSendTest extends WP_Ajax_UnitTestCase {
         $_POST = $this->create_valid_post_data([['type' => 'text', 'message' => ['text' => 'Test']]]);
 
         $response = $this->handle_ajax_request();
-        
+
         $this->assertFalse($response['success']);
         $this->assertArrayHasKey('error', $response['data']);
     }
 
     public function test_missing_required_parameters() {
         $this->_setRole('administrator');
-        
+
         // channelなし
         $_POST = $this->create_valid_post_data([['type' => 'text', 'message' => ['text' => 'Test']]]);
         unset($_POST['channel']);
@@ -87,7 +96,7 @@ class ChatSendTest extends WP_Ajax_UnitTestCase {
         $_POST['channel'] = 'invalid_channel';
 
         $response = $this->handle_ajax_request();
-        
+
         $this->assertFalse($response['success']);
         $this->assertEquals('Channel not found.', $response['data']['error'][0]);
     }
@@ -97,7 +106,7 @@ class ChatSendTest extends WP_Ajax_UnitTestCase {
         $_POST = $this->create_valid_post_data('invalid_messages_format');
 
         $response = $this->handle_ajax_request();
-        
+
         $this->assertFalse($response['success']);
         $this->assertEquals('Channel or User is not set.', $response['data']['error'][0]);
     }
