+++ tests/Publish/SavePostTest.php
@@ -9,13 +9,25 @@
 use Shipweb\LineConnect\Publish\Post as PublishPost;
 use Shipweb\LineConnect\Core\LineConnect;
 
+require_once __DIR__ . '/../../tests/LINEBot/Util/DummyHttpClient.php';
+
 class SavePostTest extends WP_UnitTestCase {
     protected static $post_id;
 
     public static function wpSetUpBeforeClass($factory) {
         add_action('save_post', array(PublishPost::class, 'save_post'), 10, 2);
-
         lineconnectTest::init();
+    }
+
+    public function setUp(): void {
+        parent::setUp();
+        add_filter(LineConnect::FILTER_PREFIX . 'httpclient', function ($httpClient) {
+            $mock = function ($testRunner, $httpMethod, $url, $data) {
+                return ['status' => 200];
+            };
+            $dummyHttpClient = new LINE\Tests\LINEBot\Util\DummyHttpClient($this, $mock);
+            return $dummyHttpClient;
+        });
     }
 
     public function test_SavePostSendNotification() {
